requirements.txt
requests
pandas
numpy
python-dotenv
solana
solders
